requirements.txt
mkdocs==1.6.1
PyYAML>=6,<7
